Restructure project: Move server-side files to a dedicated 'server' directory
